change user input from int to float when we have a list of perecentages@

+++ mdss/parser.py
@@ -205,7 +205,7 @@ def parse_args(arg_list):
     require_property_argument(gp.Angles, "atom_selection")
 
     if "," in args.size:
-        args.size = [int(s) for s in args.size.split(",")]
+        args.size = [float(s) for s in args.size.split(",")]
 
     print(args)
